AccountManager: Refactor, move db code into AccountDbHelper Make access to database consistent with how it is done in other DbHelper classes in the project. Do not open and close the database for each method call.
